# git sVarious optimizations, 4 files to go Scala.js.