… benchmarks/src/resources/scala-js/Primality.scala
… benchmarks/src/resources/scala-js/ScalaJSClassEmitter.scala
… core/src/test/resources/standard/Advanced.stat
… core/src/test/resources/standard/Apply.stat
… core/src/test/resources/standard/Val.stat
… core/src/test/resources/unit/Advanced.source
… core/src/test/resources/unit/Argument.source
… core/src/test/resources/unit/Comment.stat
… core/src/test/resources/unit/DefDef.stat
… core/src/test/resources/unit/TermApply.stat
… core/src/test/resources/unit/TypeArgument.source
… core/src/test/scala/org/scalafmt/FormatTests.scala
… core/src/test/scala/org/scalafmt/UnitTests.scala
… core/src/test/scala/org/scalafmt/util/ScalaProjectsExperiment.scala
… core/src/test/scala/org/scalafmt/util/Speed.scala
… readme.md