* Ignore singleline if obviously won't work.
* Refactor optimizations, enable by flag.
* Recurse on blocks inside no-opt zones, blocks are memoized.
* Temporary fix for comments exceeding column limit.
* Add a bunch of tests in Advanced.stat.
* Early eliminate competing solutions if there is a 0 cost solution.

```
[info] Benchmark                        Mode  Cnt     Score    Error  Units
[info] GenJsCode.scalafmt               avgt   10  1283.002 ± 77.631  ms/op
[info] GenJsCode.scalariform            avgt   10   228.591 ± 10.815  ms/op
[info] OptimizerCore.scalafmt           avgt   10  1242.236 ± 46.238  ms/op
[info] OptimizerCore.scalariform        avgt   10   230.396 ±  8.253  ms/op
[info] ScalaJSClassEmitter.scalafmt     avgt   10   308.831 ± 15.678  ms/op
[info] ScalaJSClassEmitter.scalariform  avgt   10    35.309 ±  0.602  ms/op

Success: 909
Total: 920

+-------------+--------+--------------+----------+--------+---------+----------+
|          Max|     Min|           Sum|      Mean|      Q1|       Q2|        Q3|
+-------------+--------+--------------+----------+--------+---------+----------+
|15.362,715 ms|2,165 ms|309.145,298 ms|340,094 ms|29,87 ms|74,412 ms|197,866 ms|
+-------------+--------+--------------+----------+--------+---------+----------+
```

NOTE. Only 4 files time out when formatted individually, the other 7
files probably suffer from the parallel run.